COH-556 Document key is just subject rather than subject and graph.
As the graph uri is no longer in the key of the document, it is added
as a field to the document. With the graph uri no longer part of the key
there can be multiple graph uri's added to the document. Each graph uri is
only added once.

Also updated jena arq dependency.

Change-Id: I8e5ab7bf2cb49c1ba74a68a6b874df13b7a00e7a